% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_pubnet.R
\name{add_to_pubnet}
\alias{add_to_pubnet}
\title{Function to add a new matrix of coauthorship to an existing matrix}
\usage{
add_to_pubnet(newpub, data, shape = c("symmetric", "lower"))
}
\arguments{
\item{newpub}{matrix of coauthorships to be integrated into the existing matrix.}

\item{data}{the existing matrix of coauthors<hip.}

\item{shape}{specify whether the matrix being imported is "symmetric" or "lower" triangle. Default is "symmetric".}
}
\value{
a new matrix that combines the original data matrix with the new publication matrix.
}
\description{
This functions integrates a matrix of coauthorship from a new publication into an existing matrix.
}
