% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coauthor_list.R
\name{coauthor_list}
\alias{coauthor_list}
\title{Function to get a coauthor list from a reference list}
\usage{
coauthor_list(reflist, sort = c("default", "author", "pubnumber"))
}
\arguments{
\item{reflist}{reference list to break down, imported via synthesisr::read_refs.}

\item{sort}{how to sort the data.frame. Options: "default" to sort by the reflist authors (required for network analysis),
"author" to sort by surname, or "pubnumber" to sort by the number of occurrences.}
}
\value{
a data.frame of authors and the number of occurrences in 'reflist'.
}
\description{
Gets a list of all coauthors and the number of occurrences, from a bibliography imported via the
"synthesisr" package.
}
