% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coauthor_network.R
\name{coauthor_network}
\alias{coauthor_network}
\title{Function to produce a network from a reflist}
\usage{
coauthor_network(reflist)
}
\arguments{
\item{reflist}{reference list to break down, imported via synthesisr::read_refs.}
}
\value{
A list comprising two elements: \code{network}, a tidygraph tbl_graph that can be used for plotting purposes;
and \code{coauthored}, a tibble of author occurence in reflist.
}
\description{
Core function to produce a network and a coauthorship list from a network of coauthorships.
The function internally calls \code{\link[=get_coauthor_dataframe]{get_coauthor_dataframe()}}, which flags any publications that are
single-authorships, and removes these from the dataset; and \code{\link[=coauthor_list]{coauthor_list()}}, which gives the
total number of occurrences of each author in the reference list.
}
