% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSymm.R
\name{makeSymm}
\alias{makeSymm}
\title{Function to make a lower triangle matrix a symmetric matrix}
\usage{
makeSymm(m)
}
\arguments{
\item{m}{lower triangle matrix}
}
\value{
a new matrix that combines the original data matrix with the new publication matrix.
}
\description{
This functions converts a lower triangle matrix into a symmetric matrix by transposing the lower triangle into the upper triangle.
It is taken from https://stackoverflow.com/questions/33026183/r-make-symmetric-matrix-from-lower-diagonal
}
